+++ CHANGELOG.md
@@ -21,6 +21,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 - New filter `Predicate`, `In`, that returns true if a resource is in
   a given manifest, uniquely identified by GVK, namespace, and name
   [#50](https://github.com/manifestival/manifestival/pull/50)
+- New filter `Predicate`, `ByAnnotation`, that does for annotations
+  what `ByLabel` did for labels!
+  [#52](https://github.com/manifestival/manifestival/pull/52)
 
 ### Removed